requirements.txt
requests
websocket-client
python-telegram-bot
loguru
paho-mqtt
python-logging-loki
dynaconf
pygobject